% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhsa_api_functions.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  theme = NULL,
  sub_theme = NULL,
  topic = NULL,
  geography_type = NULL,
  geography = NULL,
  metric = NULL
)
}
\arguments{
\item{theme}{the largest overall topical subgroup of data, for example \code{infectious_disease}.}

\item{sub_theme}{a topical subgroup associated with the parent theme, for example \code{respiratory}.}

\item{topic}{categorical subgroup associated with the selected theme and sub_theme. For example a topic
of \code{COVID-19} would only be available for a theme of \code{infectious_disease} and a sub_theme of
\code{respiratory}.}

\item{geography_type}{the overarching area type for the intended geography, for example \code{Nation}.}

\item{geography}{the selected area under the \code{geography_type}, for example \code{England}.}

\item{metric}{the type of data being selected, for example \code{COVID-19_testing_PCRcountByDay}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      List of parameter options.
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{      Tibble of data for the given \code{parameters}.
}\if{html}{\out{</div>}}
}
\description{
Interatively runs query to download all pages, combine the results and return a tibble.
}
\details{
For further information on the UKHSA dashboard API please visit
the \href{https://ukhsa-dashboard.data.gov.uk/access-our-data}{API documentation}.
}
\examples{
data <- get_data(
          theme = "infectious_disease",
          sub_theme = "respiratory",
          topic = "COVID-19",
          geography_type = "Nation",
          geography = "England",
          metric = "COVID-19_cases_casesByDay"
          )
}
